refactor: rename ExplorerDrawer filename
